Ticket #1918: viewer in QuickView mode corrupts screen.
Initial step: code refactoring:
  * small code formatting;
  * type accuracy;
  * introduce new type 'panel_view_mode_t' for panel type.
  * MC Viewer API unification: arguments order was changed.
  * Fixed show of viewer status bar.
  * Fixed boundaries of display text area.
  * Minor optimization of MC viewer. Call mcview_compute_areas() not so often.
  * Show not-printable 8-bit symbols in non-UTF-8 locales as dots.

+++ src/boxes.c
@@ -218,13 +218,14 @@ display_init (int radio_sel, char *init_text, int _check_status,
     return dd;
 }
 
+/* return list type */
 int
 display_box (WPanel *panel, char **userp, char **minip, int *use_msformat, int num)
 {
     int result = -1;
     Dlg_head *dd;
     char *section = NULL;
-    int i;
+    size_t i;
 
     if (panel == NULL) {
         const char *p = get_nth_panel_name (num);
